[#14] refact: Repository 구현체 분리
