--- Persistence/Persistence/Persistence.swift
@@ -1,9 +1,0 @@
-//
-//  Persistence.swift
-//  Persistence
-//
-//  Created by 문종식 on 6/12/26.
-//
-
-import Foundation
-
