+++ Repositories/Repositories/DataMapping/Networking/Response/Feedback/FeedbackStateResponse+Extension.swift
