+++ Repositories/Repositories/DataMapping/Networking/Response/Fortune/FortuneResponse+Extension.swift
