+++ Repositories/Repositories/DataMapping/Networking/Response/Record/CategoryStatisticsResponse+Extension.swift
