+++ Repositories/Repositories/DataMapping/Networking/Response/Record/MonthlyRecordResponse+Extension.swift
