+++ Repositories/Repositories/DataMapping/Networking/Response/Record/RecordResponse+Extension.swift
