+++ Repositories/Repositories/DataMapping/Networking/Response/Record/StatisticsResponse+Extension.swift
