+++ Repositories/Repositories/DataMapping/Networking/Response/Record/SummaryResponse+Extension.swift
