--- Repositories/Repositories/Repositories.swift
@@ -1,9 +1,0 @@
-//
-//  Repositories.swift
-//  Repositories
-//
-//  Created by 문종식 on 6/12/26.
-//
-
-import Foundation
-
